requirements.txt
beautifulsoup4
clint
requests
SQLAlchemy
Pillow
selenium